elapsedTime should skip manual or filtered review
